+++ packages/desktop/views/dashboard/new-dashboard/tab-section/portfolio/components/TokenListRow.svelte
@@ -1,18 +1,77 @@
 <script lang="ts">
-    import { Pane } from '@ui'
-    import { selectedAccount } from '@core/account/stores'
+    import { formatCurrency } from '@core/i18n'
+    import { getMarketAmountFromTokenValue, getMarketPriceForToken } from '@core/market/actions'
+    import { getNameFromNetworkId } from '@core/network'
+    import { ITokenWithBalance, formatTokenAmountBestMatch, getUnitFromTokenMetadata } from '@core/token'
+    import { truncateString } from '@core/utils'
+    import { PopupId, openPopup } from '@desktop/auxiliary/popup'
+    import { FontWeight, Text, TokenAvatar } from '@ui'
+
+    export let token: ITokenWithBalance
+
+    function getFormattedMarketPriceForToken(token: ITokenWithBalance): string {
+        const marketPrice = getMarketPriceForToken(token)
+        return marketPrice ? formatCurrency(marketPrice) : '-'
+    }
+
+    function getFormattedMarketPriceForTokenAmount(token: ITokenWithBalance): string {
+        const marketPrice = getMarketAmountFromTokenValue(token.balance.total, token)
+        return marketPrice ? formatCurrency(marketPrice) : '-'
+    }
+
+    function onTokenRowClick(): void {
+        openPopup({
+            id: PopupId.TokenInformation,
+            overflow: true,
+            props: {
+                token,
+            },
+        })
+    }
 </script>
 
-{#if $selectedAccount}
-    <wallet-container
-        class="w-full h-full flex flex-nowrap p-8 relative flex-1
-        bg-gray-50 dark:bg-gray-900 justify-center items-center"
+
+<button on:click={onTokenRowClick} class="token-row">
+    <div class="flex flex-row gap-4 items-start">
+        <TokenAvatar token={token} size="md" />
+        <div class="flex flex-col items-start justify-between">
+            <Text fontWeight={FontWeight.semibold}>
+                {token.metadata.name
+                    ? truncateString(token.metadata.name, 13, 0)
+                    : truncateString(token.id, 6, 7)}
+            </Text>
+            <Text secondary fontWeight={FontWeight.semibold}>
+                {getUnitFromTokenMetadata(token.metadata)}
+            </Text>
+        </div>
+    </div>
+    <Text fontWeight={FontWeight.semibold} classes="text-start"
+        >{getNameFromNetworkId(token.networkId)}</Text
     >
-        {#key $selectedAccount?.index}
-            <div class="h-full w-full flex flex-col gap-8">
-                <Pane>Balance section</Pane>
-                <Pane classes="flex-grow">Activity/Token section</Pane>
-            </div>
-        {/key}
-    </wallet-container>
-{/if}
+    <Text fontWeight={FontWeight.semibold} classes="text-start">-</Text>
+    <Text fontWeight={FontWeight.semibold} classes="text-start"
+        >{getFormattedMarketPriceForToken(token)}</Text
+    >
+    <div class="flex flex-col items-end">
+        <Text fontWeight={FontWeight.semibold} classes="text-end"
+            >{token.metadata
+                ? formatTokenAmountBestMatch(token.balance.total, token.metadata)
+                : '-'}</Text
+        >
+        <Text fontWeight={FontWeight.semibold} secondary classes="text-end"
+            >{getFormattedMarketPriceForTokenAmount(token)}</Text
+        >
+    </div>
+</button>
+
+<style lang="scss">
+    .token-row {
+        @apply w-full;
+        @apply px-5 py-4;
+        @apply border-b border-solid border-gray-100;
+        @apply hover:bg-gray-50;
+
+        @apply grid;
+        grid-template-columns: 2fr 2fr 1fr 1fr 2fr;
+    }
+</style>
